refactor(etw): separate session routing state and record decoding
